remove a load of config
