+++ config/lookoutingesterv2/config.yaml
@@ -1,7 +1,4 @@
 postgres:
-  maxOpenConns: 5
-  maxIdleConns: 25
-  connMaxLifetime: 30m
   connection:
     host: postgres
     port: 5433
@@ -9,10 +6,8 @@ postgres:
     password: psw
     dbname: postgres
     sslmode: disable
-metrics:
-  port: 9000
+metricsPort: 9000
 pulsar:
-  enabled: true
   URL: "pulsar://pulsar:6650"
   jobsetEventsTopic: "events"
   receiveTimeout: 5s
