--- internal/armada/configuration/util.go
@@ -1,11 +1,0 @@
-package configuration
-
-func (c *SchedulingConfig) GetResourceScarcity(pool string) map[string]float64 {
-	if c.PoolResourceScarcity != nil {
-		s, ok := c.PoolResourceScarcity[pool]
-		if ok {
-			return s
-		}
-	}
-	return c.ResourceScarcity
-}
